routing check

+++ llava/serve/inference.py
@@ -125,7 +125,7 @@ def answer_question(
     print('INPUT IDS SHAPE: ', input_ids)
 
     with torch.no_grad():
-        image_inputs = processor(images=image, return_tensors="pt").to("cuda")
+        image_inputs = processor(images=image, return_tensors="pt").to("cuda").half()
         outputs = vision_model(**image_inputs)
         last_hidden_state = outputs.last_hidden_state
         image_features = last_hidden_state[:, 1:]
